+++ src/cli.ts
@@ -1,5 +1,7 @@
 #!/usr/bin/env node
+import * as minimist from 'minimist';
 import { run } from './main';
 import checkPackageUpdates from './updateVersionBanner';
+
 checkPackageUpdates();
-run(process.argv.slice(2));
+run(minimist(process.argv.slice(2)));